Installer folder restored
