+++ Installer/clicSANDinstaller.nsis
@@ -1,0 +1,37 @@
+Name clicSAND
+InstallDir "C:\Program Files\clicSAND"
+LicenseData license.txt
+Page license
+Page directory
+Page instfiles
+UninstPage uninstConfirm
+UninstPage instfiles
+
+Section "-ModelRunner"
+	SetOutPath $INSTDIR
+	File /r ..\bin\Debug\*
+	WriteUninstaller $INSTDIR\uninstaller.exe
+SectionEnd
+
+Section "-CBC"
+	SetOutPath $INSTDIR\CBC
+	File /r \\icnas2.cc.ic.ac.uk\akell\downloads\Cbc-2.7.5-win64-intel11.1\Cbc-2.7.5-win64-intel11.1\*.*
+SectionEnd
+
+Section "-Templates"
+	SetOutPath $INSTDIR\Templates
+	File /r ..\Templates\*.*
+SectionEnd
+
+Section "Uninstall"
+	RMDir /r $INSTDIR\CBC
+	RMDir /r $INSTDIR\Utils
+	RMDir /r $INSTDIR\Templates
+	Delete $INSTDIR\clicSAND.exe
+	Delete $INSTDIR\clicSAND.exe.config
+	Delete $INSTDIR\clicSAND.pdb
+	Delete $INSTDIR\System.IO.Packaging.dll
+	Delete $INSTDIR\System.IO.Packaging.xml
+	Delete $INSTDIR\uninstaller.exe
+	RMDir $INSTDIR
+SectionEnd
