+++ Installer/createinstaller.bat
@@ -1,0 +1,1 @@
+makensis clicSANDinstaller.nsis
